ANDR-97: Рефактор логики регистрации флагов
Ключи регистрируются в одном месте, при необходимости можно поставить кастомные fallback значения отдельных флагов
