--- core/feature-toggle-api/src/main/java/ru/yeahub/feature_toggle_api/FeatureKeys.kt
@@ -1,9 +1,0 @@
-package ru.yeahub.feature_toggle_api
-
-object FeatureKeys {
-    val InterviewTrainer = FeatureKey(value = "interview_trainer")
-
-    val all: Set<FeatureKey> = setOf(
-        InterviewTrainer
-    )
-}
