+++ core/feature-toggle-impl/src/main/java/ru/yeahub/feature_toggle_impl/data/FeatureFlagsRepository.kt
@@ -3,6 +3,5 @@ package ru.yeahub.feature_toggle_impl.data
 import ru.yeahub.feature_toggle_api.FeatureFlagsSnapshot
 
 internal interface FeatureFlagsRepository {
-
     suspend fun getFeatureFlagsSnapshot(): FeatureFlagsSnapshot
 }
