+++ core/feature-toggle-impl/src/main/java/ru/yeahub/feature_toggle_impl/data/FeatureFlagsRepositoryImpl.kt
@@ -4,17 +4,18 @@ import ru.yeahub.feature_toggle_api.FeatureAvailability
 import ru.yeahub.feature_toggle_api.FeatureFlagsSnapshot
 import ru.yeahub.feature_toggle_api.FeatureKey
 import ru.yeahub.feature_toggle_api.FeatureKeys
-import ru.yeahub.feature_toggle_impl.resolver.FeatureKeyBackendMapper
+import ru.yeahub.feature_toggle_impl.data.remote.FeatureFlagsRemoteDataSource
 import ru.yeahub.network_api.models.GetFeatureFlagResponse
 import timber.log.Timber
 
 internal class FeatureFlagsRepositoryImpl(
-    private val remoteDataSource: FeatureFlagsRemoteDataSource,
-    private val featureKeyBackendMapper: FeatureKeyBackendMapper
+    private val remoteDataSource: FeatureFlagsRemoteDataSource
 ) : FeatureFlagsRepository {
 
+    private val featureKeysByValue = FeatureKeys.all.associateBy(FeatureKey::value)
+
     override suspend fun getFeatureFlagsSnapshot(): FeatureFlagsSnapshot {
-        val featureAvailabilityByKey = remoteDataSource.loadAndroidFeatureFlags()
+        val featureAvailabilityByKey = loadAndroidFeatureFlags()
             .mapNotNull(::mapFeatureAvailabilityByKey)
             .toMap()
 
@@ -25,6 +26,29 @@ internal class FeatureFlagsRepositoryImpl(
         )
     }
 
+    private suspend fun loadAndroidFeatureFlags(): List<GetFeatureFlagResponse> {
+        val featureFlags = mutableListOf<GetFeatureFlagResponse>()
+        var currentPage = INITIAL_PAGE
+
+        while (true) {
+            val response = remoteDataSource.getFeatureFlags(
+                page = currentPage,
+                limit = PAGE_LIMIT
+            )
+
+            featureFlags.addAll(response.data)
+
+            if (featureFlags.size >= response.total || response.data.isEmpty()) {
+                break
+            }
+
+            currentPage += 1
+        }
+
+        return featureFlags
+    }
+
+    //Логирование зарегистрированных, но не пришедших с бэкенда фичей
     private fun logMissingBackendFeatureFlags(
         featureAvailabilityByKey: Map<FeatureKey, FeatureAvailability>
     ) {
@@ -35,7 +59,7 @@ internal class FeatureFlagsRepositoryImpl(
         }
 
         val missingBackendValues = missingBackendFeatureFlags.joinToString { featureKey ->
-            featureKeyBackendMapper.getBackendValue(featureKey = featureKey)
+            featureKey.value
         }
 
         Timber.tag(FEATURE_TOGGLE_LOG_TAG)
@@ -45,9 +69,7 @@ internal class FeatureFlagsRepositoryImpl(
     private fun mapFeatureAvailabilityByKey(
         remoteFeatureFlag: GetFeatureFlagResponse
     ): Pair<FeatureKey, FeatureAvailability>? {
-        val featureKey = featureKeyBackendMapper.fromBackendValue(
-            backendValue = remoteFeatureFlag.flag
-        )
+        val featureKey = featureKeysByValue[remoteFeatureFlag.flag]
 
         if (featureKey == null) {
             Timber.tag(FEATURE_TOGGLE_LOG_TAG)
@@ -65,3 +87,5 @@ private const val FEATURE_TOGGLE_LOG_TAG = "FeatureToggle"
 private const val UNKNOWN_BACKEND_FEATURE_FLAG_LOG_MESSAGE = "Unknown backend feature flag: %s"
 private const val MISSING_BACKEND_FEATURE_FLAGS_LOG_MESSAGE =
     "Feature flags are missing in backend response: %s"
+private const val INITIAL_PAGE = 1
+private const val PAGE_LIMIT = 100
