+++ core/feature-toggle-impl/src/main/java/ru/yeahub/feature_toggle_impl/di/FeatureToggleModule.kt
@@ -3,22 +3,21 @@ package ru.yeahub.feature_toggle_impl.di
 import org.koin.core.module.dsl.singleOf
 import org.koin.dsl.module
 import ru.yeahub.feature_toggle_api.FeatureAvailabilityService
-import ru.yeahub.feature_toggle_impl.data.FeatureFlagsRemoteDataSource
 import ru.yeahub.feature_toggle_impl.data.FeatureFlagsRepository
 import ru.yeahub.feature_toggle_impl.data.FeatureFlagsRepositoryImpl
-import ru.yeahub.feature_toggle_impl.resolver.FeatureKeyBackendMapper
+import ru.yeahub.feature_toggle_impl.data.remote.FeatureFlagsRemoteDataSource
+import ru.yeahub.feature_toggle_impl.data.remote.FeatureFlagsRemoteDataSourceImpl
 import ru.yeahub.feature_toggle_impl.resolver.FeatureValueResolver
 import ru.yeahub.feature_toggle_impl.service.DefaultFeatureAvailabilityService
 
 val featureToggleModule = module {
-    singleOf(::FeatureFlagsRemoteDataSource)
-
-    singleOf(::FeatureKeyBackendMapper)
+    single<FeatureFlagsRemoteDataSource> {
+        FeatureFlagsRemoteDataSourceImpl(apiService = get())
+    }
 
     single<FeatureFlagsRepository> {
         FeatureFlagsRepositoryImpl(
-            remoteDataSource = get(),
-            featureKeyBackendMapper = get()
+            remoteDataSource = get()
         )
     }
