+++ core/feature-toggle-impl/src/main/java/ru/yeahub/feature_toggle_impl/service/DefaultFeatureAvailabilityService.kt
@@ -52,6 +52,7 @@ internal class DefaultFeatureAvailabilityService(
         }
     }
 
+    //Загрузка флагов на старте приложения
     @Suppress("TooGenericExceptionCaught")
     private suspend fun refreshFeatureFlagsSnapshot(): FeatureFlagsSnapshot? {
         var hasLoggedRefreshFailure = false
